// src/pages/hr/leavetransaction/data.d.ts
export type LeaveTransactionItem = {
  key: number;
  id: number;
  employeeCode: string;
  employee: string;
  department: string;
  leaveType: string;
  appliedTime: number;
  from: number;
  to: number;
  days: number;
  status: number;
};

export type Pagination = {
  total: number;
  pageSize: number;
  current: number;
};

export type Params = {
  status?: string;
  name?: string;
  desc?: string;
  key?: number;
  pageSize?: number;
  currentPage?: number;
  filter?: Record<string, any[]>;
  sorter?: Record<string, any>;
};